requirements.txt
alembic==1.11.1
aniso8601==9.0.1
attrs==23.1.0
bcrypt==4.0.1
blinker==1.6.2
click==8.1.6
colorama==0.4.6
flasgger==0.9.7.1
Flask==2.3.2
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
Flask-Migrate==4.0.4
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.0.5
flask-swagger-ui==4.11.1
greenlet==2.0.2
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.18.6
jsonschema-specifications==2023.7.1
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.29.0
mistune==3.0.1
packaging==23.1
psycopg2==2.9.6
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.1
rpds-py==0.9.2
six==1.16.0
SQLAlchemy==2.0.19
typing_extensions==4.7.1
validate-email==1.3
Werkzeug==2.3.6